# app/sync_server/docs/swagger.yaml
definitions:
  controllers.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      password:
        example: "1234"
        type: string
      username:
        example: admin
        type: string
    type: object
  controllers.LoginResponse:
    properties:
      token:
        example: <Header>.<Payload>.<Signature>
        type: string
    type: object
  controllers.ProfileResponse:
    properties:
      username:
        example: admin
        type: string
    type: object
  controllers.ResponseMessage:
    properties:
      message:
        example: error message
        type: string
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Return the current health status of the server
      produces:
      - application/json
      responses:
        "200":
          description: Current health status
          schema:
            $ref: '#/definitions/controllers.HealthResponse'
      summary: Get server health status
      tags:
      - health
  /login:
    post:
      consumes:
      - application/json
      description: |-
        Authenticate user and return a token
        Only "admin" with password "1234" is valid
      parameters:
      - description: Login information
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication token
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
        "400":
          description: Invalid request body | Invalid username or password
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "500":
          description: Could not generate token
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
      summary: Login with username and password
      tags:
      - auth
  /profile:
    get:
      consumes:
      - application/json
      description: Using token and return user profile information
      parameters:
      - description: <your_token_here>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile information
          schema:
            $ref: '#/definitions/controllers.ProfileResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
      summary: Get user profile
      tags:
      - auth
swagger: "2.0"